[Refactor] Rename bi_func_tools.py → bi_tools.py, fix top-level datus_bi_core imports
- Rename bi_func_tools.py to bi_tools.py to match naming convention (database.py, etc.)
- Rename test file accordingly: test_bi_func_tools.py → test_bi_tools.py
- Wrap top-level datus_bi_core imports in try/except ImportError in:
  - datus/cli/bi_dashboard.py
  - datus/tools/bi_tools/__init__.py
  - datus/tools/bi_tools/dashboard_assembler.py
- Fixes CI ModuleNotFoundError when datus-bi-core is not installed
